requirements.txt
blessings
psutil
pyyaml
